+++ CHANGELOG.md
@@ -11,12 +11,14 @@ is breaking anyways, semantic versioning is not followed.
 ### Added
 
 - Re-implement `Client::map_component` and `map_get_component`.
+- Add an `EntityRef` type to simplify interactions with entities.
 
 ### Changed
 
 - Move the `Client` struct out of `azalea-client` into `azalea`.
 - `Client::ecs` is now an `RwLock` instead of a `Mutex`.
 - `Client::component` and `entity_component` now return a mapped RwLock guard instead of cloning the component.
+- Most functions on `Client` that previously returned `Entity` now return `EntityRef` instead.
 
 ### Fixed